feat: fix expo web
