+++ apps/native/index.js
@@ -4,7 +4,7 @@ import "react-native-get-random-values";
 
 import { registerRootComponent } from "expo";
 
-import App from "./App";
+import App from "./src/App";
 
 // registerRootComponent calls AppRegistry.registerComponent('main', () => App);
 // It also ensures that whether you load the app in the Expo client or in a native build,
