+++ apps/native/package.json
@@ -1,10 +1,12 @@
 {
   "name": "native",
-  "version": "1.0.45",
+  "version": "1.0.46",
   "private": true,
   "main": "index.js",
   "scripts": {
     "ios": "expo start --ios",
+    "dev": "expo start --web",
+    "build": "npx expo export:web",
     "android": "expo start --android",
     "build-native": "expo start --no-dev --minify",
     "lint": "eslint . --ext .ts,.tsx",
@@ -13,8 +15,9 @@
   "dependencies": {
     "@effect/schema": "^0.68.0",
     "@evolu/common": "workspace:*",
-    "@evolu/common-react": "workspace:*",
+    "@evolu/react": "workspace:*",
     "@evolu/react-native": "workspace:*",
+    "@expo/metro-runtime": "~3.2.1",
     "@react-native-community/netinfo": "11.3.1",
     "babel-plugin-module-resolver": "^5.0.2",
     "crypto-browserify": "^3.12.0",
@@ -25,18 +28,28 @@
     "expo-status-bar": "~1.12.1",
     "fast-text-encoding": "^1.0.6",
     "react": "^18.3.1",
+    "react-dom": "18.2.0",
     "react-native": "^0.74.1",
     "react-native-get-random-values": "~1.11.0",
     "react-native-picker-select": "^9.1.3",
-    "stream": "^0.0.2"
+    "react-native-web": "~0.19.10",
+    "stream": "^0.0.2",
+    "vm-browserify": "^1.1.2"
   },
   "devDependencies": {
     "@babel/core": "^7.24.5",
+    "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
+    "@babel/plugin-syntax-import-attributes": "^7.24.7",
     "@babel/plugin-transform-dynamic-import": "^7.24.1",
+    "@babel/plugin-transform-modules-amd": "^7.24.7",
     "@babel/plugin-transform-private-methods": "^7.24.1",
+    "@babel/plugin-transform-private-property-in-object": "^7.24.7",
+    "@babel/preset-env": "^7.24.7",
+    "@expo/webpack-config": "~19.0.1",
     "@types/react": "~18.3.2",
     "eslint": "^8.57.0",
     "eslint-config-evolu": "workspace:*",
+    "metro-react-native-babel-preset": "^0.77.0",
     "prettier": "^3.2.5",
     "typescript": "^5.4.5"
   }
