+++ apps/native/src/styles.ts
@@ -1,0 +1,26 @@
+import { StyleSheet } from "react-native";
+
+export const appStyles = StyleSheet.create({
+  h1: {
+    fontSize: 24,
+    marginVertical: 16,
+  },
+  h2: {
+    fontSize: 18,
+    marginVertical: 16,
+  },
+  item: {
+    flexGrow: 1,
+    flexShrink: 1,
+    fontSize: 16,
+  },
+  textInput: {
+    fontSize: 18,
+    marginBottom: 16,
+  },
+  container: {
+    flex: 1,
+    padding: 16,
+    backgroundColor: "#fff",
+  },
+});
