+++ packages/evolu-common/tsconfig.json
@@ -2,7 +2,8 @@
   "extends": "@evolu/tsconfig/universal-esm.json",
   "compilerOptions": {
     "outDir": "dist",
-    "allowJs": true
+    "allowJs": true,
+    "resolveJsonModule": true
   },
   "include": ["src", "test"],
   "exclude": ["dist", "node_modules"]
