+++ packages/evolu-react-native/src/index.web.ts
@@ -1,0 +1,1 @@
+export * from "@evolu/react";
