+++ packages/evolu-react/package.json
